refactor: move more provider-needed types to tensorzero-inference-types
Move types that providers depend on from tensorzero-core to
tensorzero-inference-types, further reducing core coupling:

- EmbeddingEncodingFormat (simple enum, used by 4 embedding providers)
- mime_type_to_ext / mime_type_to_audio_format (file extension helpers)
- LazyFileExt trait (file resolution for providers)

All original import paths preserved via pub use re-exports.
