+++ crates/tensorzero-core/src/inference/types/file.rs
@@ -242,64 +242,7 @@ impl FileExt for File {
     }
 }
 
-/// Tries to convert a mime type to a file extension, picking an arbitrary extension if there are multiple
-/// extensions for the mime type.
-/// This is used when writing a file input to object storage, and when determining the file name
-/// to provide to OpenAI (which doesn't accept mime types for file input)
-pub fn mime_type_to_ext(mime_type: &MediaType) -> Result<Option<&'static str>, Error> {
-    Ok(match mime_type {
-        _ if mime_type == "image/jpeg" => Some("jpg"),
-        _ if mime_type == "image/png" => Some("png"),
-        _ if mime_type == "image/gif" => Some("gif"),
-        _ if mime_type == "application/pdf" => Some("pdf"),
-        _ if mime_type == "image/webp" => Some("webp"),
-        _ if mime_type == "text/plain" => Some("txt"),
-        _ if mime_type == "audio/midi" => Some("mid"),
-        _ if mime_type == "audio/mpeg" || mime_type == "audio/mp3" => Some("mp3"),
-        _ if mime_type == "audio/m4a" || mime_type == "audio/mp4" => Some("m4a"),
-        _ if mime_type == "audio/ogg" => Some("ogg"),
-        _ if mime_type == "audio/x-flac" || mime_type == "audio/flac" => Some("flac"),
-        _ if mime_type == "audio/x-wav"
-            || mime_type == "audio/wav"
-            || mime_type == "audio/wave" =>
-        {
-            Some("wav")
-        }
-        _ if mime_type == "audio/amr" => Some("amr"),
-        _ if mime_type == "audio/aac" || mime_type == "audio/x-aac" => Some("aac"),
-        _ if mime_type == "audio/x-aiff" || mime_type == "audio/aiff" => Some("aiff"),
-        _ if mime_type == "audio/x-dsf" => Some("dsf"),
-        _ if mime_type == "audio/x-ape" => Some("ape"),
-        _ if mime_type == "audio/webm" => Some("webm"),
-        _ => {
-            let guess = mime_guess::get_mime_extensions_str(mime_type.as_ref())
-                .and_then(|types| types.last());
-            if guess.is_some() {
-                tracing::warn!(
-                    "Guessed file extension `{guess:?}` for MIME type `{mime_type}`. This may not be correct."
-                );
-            }
-            guess.copied()
-        }
-    })
-}
-
-/// Converts audio MIME types to OpenAI's audio format strings.
-pub fn mime_type_to_audio_format(mime_type: &MediaType) -> Result<&'static str, Error> {
-    if mime_type.type_() != mime::AUDIO {
-        return Err(Error::new(ErrorDetails::InvalidMessage {
-            message: format!("Expected audio MIME type, got: {mime_type}"),
-        }));
-    }
-
-    mime_type_to_ext(mime_type)?.ok_or_else(|| {
-        Error::new(ErrorDetails::InvalidMessage {
-            message: format!(
-                "Unsupported audio MIME type: {mime_type}. Supported types: audio/midi, audio/mpeg, audio/m4a, audio/mp4, audio/ogg, audio/x-flac, audio/x-wav, audio/amr, audio/aac, audio/x-aiff, audio/x-dsf, audio/x-ape. Please open a feature request if your provider supports another audio format: https://github.com/tensorzero/tensorzero/discussions/categories/feature-requests"
-            ),
-        })
-    })
-}
+pub use tensorzero_inference_types::{mime_type_to_audio_format, mime_type_to_ext};
 
 #[cfg(test)]
 mod tests {
